rename: api 폴더 구조 변경 및 파일 이름 변경
