+++ src/apis/customAxios.ts
@@ -1,6 +1,6 @@
-import axios from 'axios';
+import axios, { AxiosInstance } from 'axios';
 
-export const axiosCustom = axios.create({
+export const customAxios: AxiosInstance = axios.create({
   baseURL:
     process.env.NODE_ENV === 'development'
       ? 'http://localhost:3000/api/v1'
