+++ src/apis/fetchTaskData.ts
@@ -1,4 +1,4 @@
-import { axiosCustom } from './createAxios';
+import { axiosCustom } from './customAxios';
 
 export async function fetchTaskData(year: number, month: number) {
   return await axiosCustom
