--- src/apis/postBannerInfo.ts
@@ -1,7 +1,0 @@
-import { axiosCustom } from './createAxios';
-
-export async function postBannerInfo(bannerInfo: any) {
-  return await axiosCustom
-    .post('banners/custom', bannerInfo)
-    .then((res) => res.data);
-}
