+++ src/apis/tasks/putRoutine.ts
@@ -1,6 +1,6 @@
 // src/apis/tasks/putRoutine.ts
 import { TaskInfo } from '@/types/routine';
-import { axiosCustom } from '../createAxios';
+import { axiosCustom } from '../customAxios';
 
 export async function putRoutine(taskData: TaskInfo) {
   return await axiosCustom
